# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Restaurant Management API",
    "description": "A Restaurant Management API built using Go",
    "version": "1.0.0",
    "contact": {
      "email": "your.email@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8001",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter the token with the Bearer prefix, e.g. 'Bearer abcde12345'"
    }
  },
  "paths": {
...
    "/restaurants": {
      "get": {
        "tags": ["restaurants"],
        "summary": "List all restaurants",
        "description": "Returns a list of restaurants",
        "operationId": "listRestaurants",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Restaurant"
                  }
                }
              },
              "required": ["title", "data"]
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": ["restaurants"],
        "summary": "Create a new restaurant",
        "description": "Adds a new restaurant to the system",
        "operationId": "createRestaurant",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "restaurant",
            "description": "Restaurant object that needs to be created",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RestaurantInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Restaurant created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Success"
                },
                "data": {
                  "$ref": "#/definitions/Restaurant"
                }
              },
              "required": ["title", "data"]
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Restaurant": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "cuisine_type": {
          "type": "string"
        },
        "rating": {
          "type": "number",
          "format": "float",
          "minimum": 0,
          "maximum": 5
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      },
      "required": ["id", "name", "address"]
    },
    "RestaurantInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "minLength": 2,
          "maxLength": 100
        },
        "description": {
          "type": "string",
          "maxLength": 500
        },
        "address": {
          "type": "string",
          "minLength": 5,
          "maxLength": 200
        },
        "cuisine_type": {
          "type": "string"
        }
      },
      "required": ["name", "address"]
    },
    "Error": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      },
      "required": ["title", "message"]
    }
  },
...
}